Extract serial API
